Restructure project to standard src/ layout
Move package to src/charge_device_simulator/ to fix pytest import errors
and follow Python packaging best practices:

- Move device/, runtime/, model/ into src/charge_device_simulator/
- Add missing __init__.py files for ocpp_j/ and ocpp_s/
- Update setup.py with package_dir and find_packages(where="src")
- Convert all internal imports to relative imports
- Update test imports to use charge_device_simulator.* prefix
- Update CI workflow coverage path
