+++ .github/workflows/test.yml
@@ -23,6 +23,7 @@ jobs:
           python -m pip install --upgrade pip
           pip install -r requirements.txt
           pip install pytest-cov pytest-asyncio
+          pip install -e .
 
       - name: Run tests with coverage
-        run: pytest --cov=device --cov-report=term-missing -v
+        run: pytest tests/ --cov=charge_device_simulator --cov-report=term-missing -v
