+++ src/charge_device_simulator/device/error_reasons.py
