+++ src/charge_device_simulator/device/frequent_flow_options.py
