+++ src/charge_device_simulator/device/ocpp_j/device_ocpp_j201.py
@@ -3,8 +3,8 @@ import sys
 import uuid
 
 import aioconsole
-from device.ocpp_j.abstract_device_ocpp_j import AbstractDeviceOcppJ
-from device.error_reasons import ErrorReasons
+from .abstract_device_ocpp_j import AbstractDeviceOcppJ
+from ..error_reasons import ErrorReasons
 
 if sys.platform != "win32":
     # Fake call to readline module to make sure it is loaded
