+++ src/charge_device_simulator/device/utility.py
