+++ src/charge_device_simulator/model/error_message.py
